HW 17.2

+++ src/features/users/api/users.controller.ts
@@ -15,8 +15,9 @@ import { UserInputDto, UserInputQueryDto } from './dto/input/user.input.dto';
 import { UsersQueryRepository } from '../infrastructure/users.query.repository';
 import { sortNPagingUserQuery } from '../../../infrastructure/utils/query.mappers';
 import { BasicAuthGuard } from '../../../infrastructure/guards/basic.auth.guard';
+import { isUUID } from 'class-validator';
 
-@Controller('users')
+@Controller('sa/users')
 @UseGuards(BasicAuthGuard)
 export class UsersController {
   constructor(
@@ -37,14 +38,14 @@ export class UsersController {
   async createUser(@Body() userInputDto: UserInputDto) {
     const newUser = await this.usersService.createSuperUser(userInputDto);
     return newUser;
-    /*if (!newUser) {
-      throw new Error('400');
-    }*/
   }
 
   @Delete(':userId')
   @HttpCode(204)
   async deleteUser(@Param('userId') userId: string) {
+    if (!isUUID(userId)) {
+      throw new NotFoundException();
+    }
     const isDelete: boolean = await this.usersService.deleteUser(userId);
     if (!isDelete) {
       throw new NotFoundException();
